refactor: folder structure
- added renamed _forms to _components and placed forms folder inside
-made _hooks folder for (login)
